requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10.3
django-custom-user==0.6
django-extensions==1.7.4
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
geopy==1.11.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.7
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
